+++ sigla-backend/src/main/resources/bootstrap/db.changelog-5.0.0.xml
@@ -5278,4 +5278,20 @@ IVAS = IVA ACQUISTO SPLIT; GEN = GENERICO; RIC = RICAVO; IVAA = IVA ACQUISTO; CR
         </insert>
     </changeSet>
     <include file="bootstrap/drop_and_recreate_table_movimenti_mag.xml"/>
+
+    <changeSet id="alter-table-pk-obbligazione-pluriennale" author="valentina.tomada">
+
+        <dropPrimaryKey tableName="obbligazione_pluriennale"></dropPrimaryKey>
+        <dropColumn tableName="obbligazione_pluriennale" columnName="id_obbligazione_plur"></dropColumn>
+        <addPrimaryKey tableName="obbligazione_pluriennale" constraintName="obbligazione_pluriennale_pkey" columnNames="cd_cds_rif,esercizio_rif,esercizio_originale_rif,pg_obbligazione_rif,anno"></addPrimaryKey>
+
+    </changeSet>
+    <changeSet id="alter-table-pk-obbligazione-pluriennale02" author="valentina.tomada">
+
+        <dropPrimaryKey tableName="obbligazione_pluriennale"></dropPrimaryKey>
+        <addPrimaryKey tableName="obbligazione_pluriennale" constraintName="obbligazione_pluriennale_pkey" columnNames="cd_cds,esercizio,esercizio_originale,pg_obbligazione,anno"></addPrimaryKey>
+
+    </changeSet>
+
+
 </databaseChangeLog>
