+++ sigla-ejb/src/main/java/it/cnr/contab/doccont00/core/bulk/Obbligazione_pluriennaleBase.java
@@ -1,25 +1,10 @@
 /*
  * Created by BulkGenerator 2.0 [07/12/2009]
- * Date 15/09/2021
+ * Date 20/09/2021
  */
 package it.cnr.contab.doccont00.core.bulk;
 import it.cnr.jada.persistency.Keyed;
 public class Obbligazione_pluriennaleBase extends Obbligazione_pluriennaleKey implements Keyed {
-//    ANNO DECIMAL(38,0) NOT NULL
-	private Integer anno;
- 
-//    CD_CDS VARCHAR(30) NOT NULL
-	private String cdCds;
- 
-//    ESERCIZIO DECIMAL(38,0) NOT NULL
-	private Integer esercizio;
- 
-//    ESERCIZIO_ORIGINALE DECIMAL(38,0) NOT NULL
-	private Integer esercizioOriginale;
- 
-//    PG_OBBLIGAZIONE DECIMAL(38,0) NOT NULL
-	private Long pgObbligazione;
- 
 //    IMPORTO DECIMAL(15,2)
 	private java.math.BigDecimal importo;
  
@@ -42,78 +27,8 @@ public class Obbligazione_pluriennaleBase extends Obbligazione_pluriennaleKey im
 	public Obbligazione_pluriennaleBase() {
 		super();
 	}
-	public Obbligazione_pluriennaleBase(Long idObbligazionePlur) {
-		super(idObbligazionePlur);
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Restituisce il valore di: [Anno Obbligazione Pluriennale]
-	 **/
-	public Integer getAnno() {
-		return anno;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Setta il valore di: [Anno Obbligazione Pluriennale]
-	 **/
-	public void setAnno(Integer anno)  {
-		this.anno=anno;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Restituisce il valore di: [Cds dell'obbligazione]
-	 **/
-	public String getCdCds() {
-		return cdCds;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Setta il valore di: [Cds dell'obbligazione]
-	 **/
-	public void setCdCds(String cdCds)  {
-		this.cdCds=cdCds;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Restituisce il valore di: [Esercizio dell'obbligazione]
-	 **/
-	public Integer getEsercizio() {
-		return esercizio;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Setta il valore di: [Esercizio dell'obbligazione]
-	 **/
-	public void setEsercizio(Integer esercizio)  {
-		this.esercizio=esercizio;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Restituisce il valore di: [Esercizio Originale dell'obbligazione]
-	 **/
-	public Integer getEsercizioOriginale() {
-		return esercizioOriginale;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Setta il valore di: [Esercizio Originale dell'obbligazione]
-	 **/
-	public void setEsercizioOriginale(Integer esercizioOriginale)  {
-		this.esercizioOriginale=esercizioOriginale;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Restituisce il valore di: [Numero dell'obbligazione]
-	 **/
-	public Long getPgObbligazione() {
-		return pgObbligazione;
-	}
-	/**
-	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Setta il valore di: [Numero dell'obbligazione]
-	 **/
-	public void setPgObbligazione(Long pgObbligazione)  {
-		this.pgObbligazione=pgObbligazione;
+	public Obbligazione_pluriennaleBase(String cdCds, Integer esercizio, Integer esercizioOriginale, Long pgObbligazione, Integer anno) {
+		super(cdCds, esercizio, esercizioOriginale, pgObbligazione, anno);
 	}
 	/**
 	 * Created by BulkGenerator 2.0 [07/12/2009]
