+++ sigla-ejb/src/main/java/it/cnr/contab/doccont00/core/bulk/Obbligazione_pluriennaleKey.java
@@ -1,12 +1,16 @@
 /*
  * Created by BulkGenerator 2.0 [07/12/2009]
- * Date 15/09/2021
+ * Date 20/09/2021
  */
 package it.cnr.contab.doccont00.core.bulk;
 import it.cnr.jada.bulk.OggettoBulk;
 import it.cnr.jada.persistency.KeyedPersistent;
 public class Obbligazione_pluriennaleKey extends OggettoBulk implements KeyedPersistent {
-	private Long idObbligazionePlur;
+	private String cdCds;
+	private Integer esercizio;
+	private Integer esercizioOriginale;
+	private Long pgObbligazione;
+	private Integer anno;
 	/**
 	 * Created by BulkGenerator 2.0 [07/12/2009]
 	 * Table name: OBBLIGAZIONE_PLURIENNALE
@@ -14,34 +18,102 @@ public class Obbligazione_pluriennaleKey extends OggettoBulk implements KeyedPer
 	public Obbligazione_pluriennaleKey() {
 		super();
 	}
-	public Obbligazione_pluriennaleKey(Long idObbligazionePlur) {
+	public Obbligazione_pluriennaleKey(String cdCds, Integer esercizio, Integer esercizioOriginale, Long pgObbligazione, Integer anno) {
 		super();
-		this.idObbligazionePlur=idObbligazionePlur;
+		this.cdCds=cdCds;
+		this.esercizio=esercizio;
+		this.esercizioOriginale=esercizioOriginale;
+		this.pgObbligazione=pgObbligazione;
+		this.anno=anno;
 	}
 	public boolean equalsByPrimaryKey(Object o) {
 		if (this== o) return true;
 		if (!(o instanceof Obbligazione_pluriennaleKey)) return false;
 		Obbligazione_pluriennaleKey k = (Obbligazione_pluriennaleKey) o;
-		if (!compareKey(getIdObbligazionePlur(), k.getIdObbligazionePlur())) return false;
+		if (!compareKey(getCdCds(), k.getCdCds())) return false;
+		if (!compareKey(getEsercizio(), k.getEsercizio())) return false;
+		if (!compareKey(getEsercizioOriginale(), k.getEsercizioOriginale())) return false;
+		if (!compareKey(getPgObbligazione(), k.getPgObbligazione())) return false;
+		if (!compareKey(getAnno(), k.getAnno())) return false;
 		return true;
 	}
 	public int primaryKeyHashCode() {
 		int i = 0;
-		i = i + calculateKeyHashCode(getIdObbligazionePlur());
+		i = i + calculateKeyHashCode(getCdCds());
+		i = i + calculateKeyHashCode(getEsercizio());
+		i = i + calculateKeyHashCode(getEsercizioOriginale());
+		i = i + calculateKeyHashCode(getPgObbligazione());
+		i = i + calculateKeyHashCode(getAnno());
 		return i;
 	}
 	/**
 	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Restituisce il valore di: [Chiave Primaria.]
+	 * Restituisce il valore di: [Cds dell'obbligazione]
 	 **/
-	public void setIdObbligazionePlur(Long idObbligazionePlur)  {
-		this.idObbligazionePlur=idObbligazionePlur;
+	public void setCdCds(String cdCds)  {
+		this.cdCds=cdCds;
 	}
 	/**
 	 * Created by BulkGenerator 2.0 [07/12/2009]
-	 * Setta il valore di: [Chiave Primaria.]
+	 * Setta il valore di: [Cds dell'obbligazione]
 	 **/
-	public Long getIdObbligazionePlur() {
-		return idObbligazionePlur;
+	public String getCdCds() {
+		return cdCds;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Restituisce il valore di: [Esercizio dell'obbligazione]
+	 **/
+	public void setEsercizio(Integer esercizio)  {
+		this.esercizio=esercizio;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Setta il valore di: [Esercizio dell'obbligazione]
+	 **/
+	public Integer getEsercizio() {
+		return esercizio;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Restituisce il valore di: [Esercizio Originale dell'obbligazione]
+	 **/
+	public void setEsercizioOriginale(Integer esercizioOriginale)  {
+		this.esercizioOriginale=esercizioOriginale;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Setta il valore di: [Esercizio Originale dell'obbligazione]
+	 **/
+	public Integer getEsercizioOriginale() {
+		return esercizioOriginale;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Restituisce il valore di: [Numero dell'obbligazione]
+	 **/
+	public void setPgObbligazione(Long pgObbligazione)  {
+		this.pgObbligazione=pgObbligazione;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Setta il valore di: [Numero dell'obbligazione]
+	 **/
+	public Long getPgObbligazione() {
+		return pgObbligazione;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Restituisce il valore di: [Anno Obbligazione Pluriennale]
+	 **/
+	public void setAnno(Integer anno)  {
+		this.anno=anno;
+	}
+	/**
+	 * Created by BulkGenerator 2.0 [07/12/2009]
+	 * Setta il valore di: [Anno Obbligazione Pluriennale]
+	 **/
+	public Integer getAnno() {
+		return anno;
 	}
 }
