+++ sigla-ejb/src/main/resources/it/cnr/contab/doccont00/core/bulk/Obbligazione_pluriennaleBasePersistentInfo.xml
@@ -6,36 +6,6 @@
 	<defaultColumnMap
 		tableName="OBBLIGAZIONE_PLURIENNALE">
 		<columnMapping
-			columnName="ANNO"
-			propertyName="anno"
-			sqlTypeName="DECIMAL"
-			columnSize="38"
-			nullable="false" />
-		<columnMapping
-			columnName="CD_CDS"
-			propertyName="cdCds"
-			sqlTypeName="VARCHAR"
-			columnSize="30"
-			nullable="false" />
-		<columnMapping
-			columnName="ESERCIZIO"
-			propertyName="esercizio"
-			sqlTypeName="DECIMAL"
-			columnSize="38"
-			nullable="false" />
-		<columnMapping
-			columnName="ESERCIZIO_ORIGINALE"
-			propertyName="esercizioOriginale"
-			sqlTypeName="DECIMAL"
-			columnSize="38"
-			nullable="false" />
-		<columnMapping
-			columnName="PG_OBBLIGAZIONE"
-			propertyName="pgObbligazione"
-			sqlTypeName="DECIMAL"
-			columnSize="38"
-			nullable="false" />
-		<columnMapping
 			columnName="IMPORTO"
 			propertyName="importo"
 			sqlTypeName="DECIMAL"
@@ -97,16 +67,6 @@
 			columnSize="38"
 			nullable="false" />
 	</defaultColumnMap>
-	<persistentProperty
-		name="anno" />
-	<persistentProperty
-		name="cdCds" />
-	<persistentProperty
-		name="esercizio" />
-	<persistentProperty
-		name="esercizioOriginale" />
-	<persistentProperty
-		name="pgObbligazione" />
 	<persistentProperty
 		name="importo" />
 	<persistentProperty
