+++ sigla-ejb/src/main/resources/it/cnr/contab/doccont00/core/bulk/Obbligazione_pluriennaleKeyPersistentInfo.xml
@@ -4,14 +4,54 @@
 	<defaultColumnMap
 		tableName="OBBLIGAZIONE_PLURIENNALE">
 		<columnMapping
-			columnName="ID_OBBLIGAZIONE_PLUR"
-			propertyName="idObbligazionePlur"
+			columnName="CD_CDS"
+			propertyName="cdCds"
+			sqlTypeName="VARCHAR"
+			columnSize="30"
+			nullable="false"
+			primary="true" />
+		<columnMapping
+			columnName="ESERCIZIO"
+			propertyName="esercizio"
+			sqlTypeName="DECIMAL"
+			columnSize="38"
+			nullable="false"
+			primary="true" />
+		<columnMapping
+			columnName="ESERCIZIO_ORIGINALE"
+			propertyName="esercizioOriginale"
+			sqlTypeName="DECIMAL"
+			columnSize="38"
+			nullable="false"
+			primary="true" />
+		<columnMapping
+			columnName="PG_OBBLIGAZIONE"
+			propertyName="pgObbligazione"
+			sqlTypeName="DECIMAL"
+			columnSize="38"
+			nullable="false"
+			primary="true" />
+		<columnMapping
+			columnName="ANNO"
+			propertyName="anno"
 			sqlTypeName="DECIMAL"
 			columnSize="38"
 			nullable="false"
 			primary="true" />
 	</defaultColumnMap>
 	<persistentProperty
-		name="idObbligazionePlur"
+		name="cdCds"
+		partOfOid="true" />
+	<persistentProperty
+		name="esercizio"
+		partOfOid="true" />
+	<persistentProperty
+		name="esercizioOriginale"
+		partOfOid="true" />
+	<persistentProperty
+		name="pgObbligazione"
+		partOfOid="true" />
+	<persistentProperty
+		name="anno"
 		partOfOid="true" />
 </SQLPersistentInfo>
